Bug Fix:Monkey Patch Sitemap Generator is lib not config
